requirements.txt
urllib3~=1.26.3
python-dotenv~=0.17.0
cryptography